+++ packages/orq-rc/RELEASES.md
@@ -4539,3 +4539,13 @@ Based on:
 - [typescript v3.14.0-rc.54] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.54] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.54 - packages/orq-rc
+
+## 2025-10-24 09:50:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.639.1 (2.730.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.55] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.55] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.55 - packages/orq-rc